+++ src/consumer/component-ops/import-components.js
@@ -2,7 +2,7 @@
 import R from 'ramda';
 import chalk from 'chalk';
 import { NothingToImport } from '../exceptions';
-import { BitId } from '../../bit-id';
+import { BitId, BitIds } from '../../bit-id';
 import Component from '../component';
 import { Consumer } from '../../consumer';
 import { ComponentWithDependencies, Scope } from '../../scope';
@@ -19,6 +19,9 @@ import ManyComponentsWriter from './many-components-writer';
 import { COMPONENT_ORIGINS } from '../../constants';
 import hasWildcard from '../../utils/string/has-wildcard';
 import { getRemoteBitIdsByWildcards } from '../../api/consumer/lib/list-scope';
+import { getScopeRemotes } from '../../scope/scope-remotes';
+import Remotes from '../../remotes/remotes';
+import DependencyGraph from '../../scope/graph/scope-graph';
 
 export type ImportOptions = {
   ids: string[], // array might be empty
@@ -34,7 +37,9 @@ export type ImportOptions = {
   override: boolean, // default: false
   installNpmPackages: boolean, // default: true
   objectsOnly: boolean, // default: false
-  saveDependenciesAsComponents?: boolean // default: false
+  saveDependenciesAsComponents?: boolean, // default: false,
+  importDependenciesDirectly?: boolean, // default: false, normally it imports them as packages or nested, not as imported
+  importDependents?: boolean // default: false,
 };
 type ComponentMergeStatus = {
   componentWithDependencies: ComponentWithDependencies,
@@ -76,7 +81,7 @@ export default class ImportComponents {
 
   async importSpecificComponents(): ImportResult {
     logger.debug(`importSpecificComponents, Ids: ${this.options.ids.join(', ')}`);
-    const bitIds = await this._getBitIds();
+    const bitIds: BitIds = await this._getBitIds();
     const beforeImportVersions = await this._getCurrentVersions(bitIds);
     await this._throwForPotentialIssues(bitIds);
     const componentsWithDependencies = await this.consumer.importComponents(
@@ -90,7 +95,7 @@ export default class ImportComponents {
     return { dependencies: componentsWithDependencies, importDetails };
   }
 
-  async _getBitIds(): Promise<BitId[]> {
+  async _getBitIds(): Promise<BitIds> {
     const bitIds: BitId[] = [];
     await Promise.all(
       this.options.ids.map(async (idStr: string) => {
@@ -103,7 +108,47 @@ export default class ImportComponents {
         }
       })
     );
-    return bitIds;
+    if (this.options.importDependenciesDirectly || this.options.importDependents) {
+      const graphs = await this._getComponentsGraphs(bitIds);
+      if (this.options.importDependenciesDirectly) {
+        const dependenciesIds = this._getDependenciesFromGraph(bitIds, graphs);
+        bitIds.push(...dependenciesIds);
+      }
+      if (this.options.importDependents) {
+        const dependentsIds = this._getDependentsFromGraph(bitIds, graphs);
+        bitIds.push(...dependentsIds);
+      }
+    }
+    return BitIds.uniqFromArray(bitIds);
+  }
+
+  _getDependenciesFromGraph(bitIds: BitId[], graphs: DependencyGraph[]): BitId[] {
+    const dependencies = bitIds.map((bitId) => {
+      const componentGraph = graphs.find(graph => graph.scopeName === bitId.scope);
+      if (!componentGraph) {
+        throw new Error(`unable to find a graph for ${bitId.toString()}`);
+      }
+      const dependenciesInfo = componentGraph.getDependenciesInfo(bitId);
+      return dependenciesInfo.map(d => d.id);
+    });
+    return R.flatten(dependencies);
+  }
+
+  _getDependentsFromGraph(bitIds: BitId[], graphs: DependencyGraph[]): BitId[] {
+    const dependents = bitIds.map((bitId) => {
+      const componentGraph = graphs.find(graph => graph.scopeName === bitId.scope);
+      if (!componentGraph) {
+        throw new Error(`unable to find a graph for ${bitId.toString()}`);
+      }
+      const dependentsInfo = componentGraph.getDependentsInfo(bitId);
+      return dependentsInfo.map(d => d.id);
+    });
+    return R.flatten(dependents);
+  }
+
+  async _getComponentsGraphs(bitIds: BitId[]): Promise<DependencyGraph[]> {
+    const remotes: Remotes = await getScopeRemotes(this.consumer.scope);
+    return remotes.scopeGraphs(bitIds, this.consumer.scope);
   }
 
   async importAccordingToBitMap(): ImportResult {
@@ -158,7 +203,7 @@ export default class ImportComponents {
     return { dependencies: componentsAndDependencies, importDetails };
   }
 
-  async _getCurrentVersions(ids: BitId[]): ImportedVersions {
+  async _getCurrentVersions(ids: BitIds): ImportedVersions {
     const versionsP = ids.map(async (id) => {
       const modelComponent = await this.consumer.scope.getModelComponentIfExist(id);
       const idStr = id.toStringWithoutVersion();
@@ -202,16 +247,17 @@ export default class ImportComponents {
     return Promise.all(detailsP);
   }
 
-  async _throwForPotentialIssues(ids: BitId[]): Promise<void> {
+  async _throwForPotentialIssues(ids: BitIds): Promise<void> {
     await this._throwForModifiedOrNewComponents(ids);
     this._throwForDifferentComponentWithSameName(ids);
   }
 
-  async _throwForModifiedOrNewComponents(ids: BitId[]): Promise<void> {
+  async _throwForModifiedOrNewComponents(ids: BitIds): Promise<void> {
     // the typical objectsOnly option is when a user cloned a project with components tagged to the source code, but
     // doesn't have the model objects. in that case, calling getComponentStatusById() may return an error as it relies
     // on the model objects when there are dependencies
     if (this.options.override || this.options.objectsOnly || this.options.merge) return;
+    // $FlowFixMe BitIds is an array
     const modifiedComponents = await filterAsync(ids, (id) => {
       return this.consumer.getComponentStatusById(id).then(status => status.modified || status.newlyCreated);
     });
@@ -240,7 +286,7 @@ export default class ImportComponents {
    * If an imported component has scope+name equals to a local name, both will have the exact same
    * hash and they'll override each other.
    */
-  _throwForDifferentComponentWithSameName(ids: BitId[]): void {
+  _throwForDifferentComponentWithSameName(ids: BitIds): void {
     ids.forEach((id: BitId) => {
       const existingId = this.consumer.getParsedIdIfExist(id.toStringWithoutVersion());
       if (existingId && !existingId.hasScope()) {